backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "clerk-sdk-python>=0.1.0",
    "fastapi>=0.124.2",
    "openai>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
